add plot dem

+++ scripts/remove_outlier.py
@@ -52,11 +52,15 @@ def filter(data):
 DIR = Path("./data/test/out/")
 # datas = None
 for file_ in glob.iglob(os.path.join(DIR,r"LOLARDR_*.AR")): # 匹配数据文件
+    if os.path.exists(f"{file_[:-3]}.AO"):
+        continue
     data = read_data(file_)
     data = filter(data)
     data[["lon","lat","alt","t1","t2"]].to_csv(f"{file_[:-3]}.AO", sep=" ", header = 0, index=0, float_format="%.7f")
     
 for file_ in glob.iglob(os.path.join(DIR,r"LOLARDR_*.DR")): # 匹配数据文件
+    if os.path.exists(f"{file_[:-3]}.DO"):
+        continue
     data = read_data(file_)
     data = filter(data)
     data[["lon","lat","alt","t1","t2"]].to_csv(f"{file_[:-3]}.DO", sep=" ", header = 0, index=0, float_format="%.7f")
